+++ Stacks/Linked-List-Stack/stack.h
@@ -3,9 +3,9 @@
 
 typedef struct stack Stack;
 
-Stack * create_element (int info);
-int push_stack (Stack ** head, int info);
-int pop_stack (Stack ** head);
+Stack * create_element (float info);
+int push_stack (Stack ** head, float info);
+float pop_stack (Stack ** head);
 int empty_stack (Stack * head);
 int size_stack (Stack * head);
 void free_stack (Stack ** head);
